% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atibble.R
\docType{class}
\name{atibble-class}
\alias{atibble-class}
\alias{atibble}
\title{atibble = analysis tibble}
\description{
atibble = analysis tibble
}
\section{Slots}{

\describe{
\item{\code{name}}{a name}

\item{\code{label}}{a label}
}}

