% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_count.R, R/v_continuous.R,
%   R/v_continuous_nonneg.R, R/v_event_time.R, R/v_binary.R, R/v_nominal.R,
%   R/v_ordered.R, R/v_character.R, R/v_rcensored.R
\name{casting}
\alias{casting}
\alias{vec_ptype2.v_count}
\alias{vec_cast.v_count}
\alias{vec_ptype2.v_continuous}
\alias{vec_cast.v_continuous}
\alias{vec_ptype2.v_continuous_nonneg}
\alias{vec_cast.v_continuous_nonneg}
\alias{vec_ptype2.v_event_time}
\alias{vec_cast.v_event_time}
\alias{vec_ptype2.v_binary}
\alias{vec_cast.v_binary}
\alias{vec_ptype2.v_nominal}
\alias{vec_cast.v_nominal}
\alias{vec_ptype2.v_ordered}
\alias{vec_cast.v_ordered}
\alias{vec_ptype2.v_character}
\alias{vec_cast.v_character}
\alias{vec_ptype2.v_rcensored}
\alias{vec_cast.v_rcensored}
\title{Casting}
\usage{
\method{vec_ptype2}{v_count}(x, y, ...)

\method{vec_cast}{v_count}(x, to, ...)

\method{vec_ptype2}{v_continuous}(x, y, ...)

\method{vec_cast}{v_continuous}(x, to, ...)

\method{vec_ptype2}{v_continuous_nonneg}(x, y, ...)

\method{vec_cast}{v_continuous_nonneg}(x, to, ...)

\method{vec_ptype2}{v_event_time}(x, y, ...)

\method{vec_cast}{v_event_time}(x, to, ...)

\method{vec_ptype2}{v_binary}(x, y, ...)

\method{vec_cast}{v_binary}(x, to, ...)

\method{vec_ptype2}{v_nominal}(x, y, ...)

\method{vec_cast}{v_nominal}(x, to, ...)

\method{vec_ptype2}{v_ordered}(x, y, ...)

\method{vec_cast}{v_ordered}(x, to, ...)

\method{vec_ptype2}{v_character}(x, y, ...)

\method{vec_cast}{v_character}(x, to, ...)

\method{vec_ptype2}{v_rcensored}(x, y, ...)

\method{vec_cast}{v_rcensored}(x, to, ...)
}
\arguments{
\item{x}{Vector types.}

\item{y}{Vector types.}

\item{...}{These dots are for future extensions and must be empty.}

\item{to}{Type to cast to. If \code{NULL}, \code{x} will be returned as is.}
}
\description{
Casting

Casting
}
