% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_internal_name}
\alias{check_internal_name}
\title{Check that a character vector has length 0 or 1}
\usage{
check_internal_name(x)
}
\arguments{
\item{x}{what to check}
}
\description{
Check that a character vector has length 0 or 1
}
\keyword{internal}
