% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_internal_names}
\alias{check_internal_names}
\title{Check that internal names are equal}
\usage{
check_internal_names(x, y)
}
\arguments{
\item{x}{a stype vector}

\item{y}{another stype vector}
}
\description{
Check that internal names are equal
}
\keyword{internal}
