% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_number_positive}
\alias{check_number_positive}
\title{Check whether a value is a positive scalar number}
\usage{
check_number_positive(x)
}
\arguments{
\item{x}{what to check}
}
\description{
Check whether a value is a positive scalar number
}
\keyword{internal}
