% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_summary_args}
\alias{check_summary_args}
\title{Checks that summary function ellipsis arguments only}
\usage{
check_summary_args(...)
}
\description{
Checks that summary function ellipsis arguments only
}
\keyword{internal}
