% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{context_get_set}
\alias{context_get_set}
\alias{get_context}
\alias{get_context,described-method}
\alias{get_context,described,described-method}
\alias{get_short_label}
\alias{get_short_label,context-method}
\alias{get_short_label,context,context-method}
\alias{get_short_label,described-method}
\alias{get_short_label,described,described-method}
\alias{get_long_label}
\alias{get_long_label,context-method}
\alias{get_long_label,context,context-method}
\alias{get_long_label,described-method}
\alias{get_long_label,described,described-method}
\alias{get_description}
\alias{get_description,context-method}
\alias{get_description,context,context-method}
\alias{get_description,described-method}
\alias{get_description,described,described-method}
\alias{get_derivation}
\alias{get_derivation,context-method}
\alias{get_derivation,context,context-method}
\alias{get_derivation,described-method}
\alias{get_derivation,described,described-method}
\alias{get_purpose}
\alias{get_purpose,context-method}
\alias{get_purpose,context,context-method}
\alias{get_purpose,described-method}
\alias{get_purpose,described,described-method}
\alias{get_security_type}
\alias{get_security_type,context-method}
\alias{get_security_type,context,context-method}
\alias{get_security_type,described-method}
\alias{get_security_type,described,described-method}
\title{Get an object's context or elements thereof}
\usage{
get_context(x)

\S4method{get_context}{described}(x)

get_short_label(x)

\S4method{get_short_label}{context}(x)

\S4method{get_short_label}{described}(x)

get_long_label(x)

\S4method{get_long_label}{context}(x)

\S4method{get_long_label}{described}(x)

get_description(x)

\S4method{get_description}{context}(x)

\S4method{get_description}{described}(x)

get_derivation(x)

\S4method{get_derivation}{context}(x)

\S4method{get_derivation}{described}(x)

get_purpose(x)

\S4method{get_purpose}{context}(x)

\S4method{get_purpose}{described}(x)

get_security_type(x)

\S4method{get_security_type}{context}(x)

\S4method{get_security_type}{described}(x)
}
\arguments{
\item{x}{a described object}
}
\description{
Get an object's context or elements thereof
}
