% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\docType{class}
\name{data_summary-class}
\alias{data_summary-class}
\alias{data_summary}
\title{A Class Subclassing From the \code{methods::namedList} Class}
\description{
This is the class of object returned by a call to \code{\link{describe}} with a stype
vector as the input. The class currently does not extend the
\code{\link[methods:BasicClasses]{methods::namedList}} class.
}
