% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{decide_auto_compute}
\alias{decide_auto_compute}
\title{In the case that both x and y not equal to the default,
return not the default;
otherwise, return the default}
\usage{
decide_auto_compute(x, y)
}
\description{
In the case that both x and y not equal to the default,
return not the default;
otherwise, return the default
}
\keyword{internal}
