% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aREADME.R
\docType{package}
\name{DenverSugar-package}
\alias{DenverSugar}
\alias{DenverSugar-package}
\title{DenverSugar: a grammar of cohort creation}
\description{
TODO: add a description
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.novisci.com/statistics/DenverSugar}
  \item Report bugs at \url{https://gitlab.novisci.com/statistics/DenverSugar/issues}
}

}
