% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\alias{describe,describable,maybeGroup,maybeWeight,maybeDescriptor-method}
\alias{describe,simpleStype,maybeGroup,maybeWeight,maybeDescriptor-method}
\alias{describe,tbl_df,missing,missing,maybeDescriptor-method}
\alias{describe,v_rcensored,maybeGroup,maybeWeight,maybeDescriptor-method}
\alias{weight}
\alias{weight,stype,weightVar,maybeDescriptor-method}
\title{Describe a variable}
\usage{
describe(x, g = NULL, w = NULL, .descriptors, ...)

\S4method{describe}{describable,maybeGroup,maybeWeight,maybeDescriptor}(x, g = NULL, w = NULL, .descriptors, ...)

\S4method{describe}{simpleStype,maybeGroup,maybeWeight,maybeDescriptor}(x, g = NULL, w = NULL, .descriptors, ...)

\S4method{describe}{tbl_df,missing,missing,maybeDescriptor}(x, g = NULL, w = NULL, .descriptors, ...)

\S4method{describe}{v_rcensored,maybeGroup,maybeWeight,maybeDescriptor}(x, g = NULL, w = NULL, .descriptors, ...)

weight(x, w, .descriptors)

\S4method{weight}{stype,weightVar,maybeDescriptor}(x, w, .descriptors)
}
\arguments{
\item{x}{a vector a values}

\item{g}{a vector a groupings (optional)}

\item{w}{a vector of weights (optional)}

\item{.descriptors}{an (optional) list of lambda functions}

\item{...}{additional arguments}
}
\description{
Describe a variable

(Re)\code{describe} a \code{stype} by weighting the vector.
}
