% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptor.R
\docType{class}
\name{descriptor-class}
\alias{descriptor-class}
\alias{descriptor}
\title{descriptor}
\description{
Extends R's \code{\link{vector-class}} to provide objects with additional information. 
The vector object must be one of the following basic classes: \code{"numeric", "double", 
"integer", "logical", "character",}, or \code{\link{factor}}, or \code{\link{Date}}.
 A variable object can used as if it was a vector.
}
\section{Slots}{

\describe{
\item{\code{f}}{an anyomymous with a signature of}

\item{\code{label}}{a character description of the descriptor}

\item{\code{display}}{a function for displaying the descriptor}
}}

