% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\docType{methods}
\name{descriptor}
\alias{descriptor}
\alias{descriptor,variable-method}
\alias{descriptor,logical-method}
\alias{descriptor,groupVar-method}
\alias{descriptor,maybeWeight-method}
\alias{descriptor,NULL-method}
\title{Descriptor}
\usage{
descriptor(x)

\S4method{descriptor}{variable}(x)

\S4method{descriptor}{logical}(x)

\S4method{descriptor}{groupVar}(x)

\S4method{descriptor}{maybeWeight}(x)

\S4method{descriptor}{`NULL`}(x)
}
\arguments{
\item{x}{a \linkS4class{variable} or vector}
}
\description{
A function that returns a list of functions to be applied to a variable
}
