% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\docType{data}
\name{descriptors}
\alias{descriptors}
\alias{standardDescriptors}
\alias{logicalDescriptors}
\alias{wlogicalDescriptors}
\alias{categoricalDescriptors}
\alias{characterDescriptors}
\alias{numericDescriptors}
\alias{wnumericDescriptors}
\alias{groupedDescriptors}
\title{Descriptors}
\format{An object of class \code{list} of length 6.}
\usage{
standardDescriptors

logicalDescriptors

wlogicalDescriptors

categoricalDescriptors

characterDescriptors

numericDescriptors

wnumericDescriptors

groupedDescriptors
}
\description{
Lists of functions to used to create a \code{\link{data_summary}} object.
}
\keyword{datasets}
\keyword{internal}
