% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{determine_mathish_codomain}
\alias{determine_mathish_codomain}
\title{Get the codomain of math/summary function for stype vectors}
\usage{
determine_mathish_codomain(x, fns)
}
\arguments{
\item{x}{a dummy stype vector}

\item{fns}{character vector of vector names}
}
\description{
Get the codomain of math/summary function for stype vectors
}
\keyword{internal}
