% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{eair}
\alias{eair}
\title{Exposure-Adjusted Incidence Rate}
\usage{
eair(x)
}
\arguments{
\item{x}{a \code{\link{v_rcensored}} vector}
}
\value{
a \code{list} containing:
\itemize{
\item eair the point estimate
\item eair_variance the variance estimate
}
}
\description{
Computes the exposure-adjusted incidence rate and variance using He et al.
(2015).
}
\references{
He X, Chen L, Lei L, Xia HA, Lee MLT (2015) A Simple Method for
Estimating Confidence Intervals for Exposure Adjusted Incidence Rate and
Its Applications to Clinical Trials. J Biom Biostat 6: 238.
doi:10.4172/2155-6180.1000238
}
