% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{get_censorinfo}
\alias{get_censorinfo}
\title{Get censoring info for a data_summary}
\usage{
get_censorinfo(x, ...)
}
\arguments{
\item{x}{a \code{\link{v_rcensored}} vector}
}
\description{
Get censoring info for a data_summary
}
\keyword{internal}
