% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\docType{methods}
\name{get_context}
\alias{get_context}
\alias{get_context,described-method}
\alias{get_context,described,context-method}
\title{Get an object's context}
\usage{
get_context(x)

\S4method{get_context}{described}(x)
}
\description{
Get an object's context
}
