% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{get_data_summaries}
\alias{get_data_summaries}
\alias{get_data_summaries,list,ANY-method}
\alias{get_data_summaries,list,character-method}
\title{Get data summaries from a list of stypes}
\usage{
get_data_summaries(x, element)

\S4method{get_data_summaries}{list,ANY}(x, element)

\S4method{get_data_summaries}{list,character}(x, element)
}
\arguments{
\item{x}{the object from which to get the \code{data_summaries}}

\item{element}{either \code{NULL} to get the full \code{data_summary}
or a length 1 \code{character} to select a particular element of the summary}
}
\description{
Get data summaries from a list of stypes
}
