% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{get_data_summary}
\alias{get_data_summary}
\alias{get_data_summary,stype,ANY-method}
\alias{get_data_summary,v_rcensored,ANY-method}
\alias{get_data_summary,stype,character-method}
\title{Get the summary from a stype variable}
\usage{
get_data_summary(x, element)

\S4method{get_data_summary}{stype,ANY}(x, element)

\S4method{get_data_summary}{v_rcensored,ANY}(x, element)

\S4method{get_data_summary}{stype,character}(x, element)
}
\arguments{
\item{x}{the object from which to get the \code{data_summary}}

\item{element}{either \code{NULL} to get the full \code{data_summary}
or a length 1 \code{character} to select a particular element of the summary}
}
\description{
Get the summary from a stype variable
}
