% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{get_internal_name}
\alias{get_internal_name}
\alias{get_internal_name,stype-method}
\title{Get the internal name from a stype variable}
\usage{
get_internal_name(x, element)

\S4method{get_internal_name}{stype}(x, element)
}
\arguments{
\item{x}{the object from which to get the \code{internal_name} attribute}
}
\description{
Get the internal name from a stype variable
}
