% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_math_codomains}
\alias{get_math_codomains}
\title{Get Math codomains}
\usage{
get_math_codomains(x)
}
\arguments{
\item{x}{a dummy stype vector}
}
\description{
Get Math codomains
}
\keyword{internal}
