% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{get_outcomeinfo}
\alias{get_outcomeinfo}
\title{Get outcome info for a data_summary}
\usage{
get_outcomeinfo(x, ...)
}
\arguments{
\item{x}{a \code{\link{v_rcensored}} vector}
}
\description{
Get outcome info for a data_summary
}
\keyword{internal}
