% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{get_persontime}
\alias{get_persontime}
\title{Get person-time info for a data_summary}
\usage{
get_persontime(x, ...)
}
\arguments{
\item{x}{a \code{\link{v_rcensored}} vector}
}
\description{
Get person-time info for a data_summary
}
\keyword{internal}
