% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\docType{methods}
\name{get_purpose}
\alias{get_purpose}
\alias{get_purpose,context-method}
\alias{get_purpose,context,context-method}
\alias{get_purpose,described-method}
\alias{get_purpose,described,described-method}
\title{Get an object's purpose}
\usage{
get_purpose(x)

\S4method{get_purpose}{context}(x)

\S4method{get_purpose}{described}(x)
}
\arguments{
\item{x}{a described object}
}
\description{
Get an object's purpose
}
