% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_summary_codomains}
\alias{get_summary_codomains}
\title{Get Summary codomains}
\usage{
get_summary_codomains(x)
}
\arguments{
\item{x}{a dummy stype vector}
}
\description{
Get Summary codomains
}
\keyword{internal}
