% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{getDescriptors}
\alias{getDescriptors}
\alias{getDescriptors,logical,maybeGroup,maybeWeight-method}
\alias{getDescriptors,categorical,maybeGroup,maybeWeight-method}
\alias{getDescriptors,character,maybeGroup,maybeWeight-method}
\alias{getDescriptors,numeric,maybeGroup,maybeWeight-method}
\alias{getDescriptors,v_rcensored,maybeGroup,maybeWeight-method}
\title{getDescriptors}
\usage{
getDescriptors(x, g, w)

\S4method{getDescriptors}{logical,maybeGroup,maybeWeight}(x, g, w)

\S4method{getDescriptors}{categorical,maybeGroup,maybeWeight}(x, g, w)

\S4method{getDescriptors}{character,maybeGroup,maybeWeight}(x, g, w)

\S4method{getDescriptors}{numeric,maybeGroup,maybeWeight}(x, g, w)

\S4method{getDescriptors}{v_rcensored,maybeGroup,maybeWeight}(x, g, w)
}
\arguments{
\item{x}{a vector of data}

\item{g}{a vector of groupings}

\item{w}{a vector of weights}
}
\description{
Returns a list of functions to be applied to a variable.
}
\keyword{internal}
