% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableClass.R, R/atibble.R
\docType{methods}
\name{glance}
\alias{glance}
\alias{glance,variable-method}
\alias{glance,atibble-method}
\title{Glance at a variable}
\usage{
glance(object)

\S4method{glance}{variable}(object)

\S4method{glance}{atibble}(object)
}
\arguments{
\item{object}{an object}
}
\description{
Glance at a variable
}
