% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\docType{methods}
\name{is_covariate}
\alias{is_covariate}
\alias{is_covariate,context-method}
\alias{is_covarite,context,context-method}
\alias{is_covariate,described-method}
\alias{is_covarite,described,context-method}
\title{Is this vector a covariate?}
\usage{
is_covariate(x)

\S4method{is_covariate}{context}(x)

\S4method{is_covariate}{described}(x)
}
\arguments{
\item{x}{what to check}
}
\description{
Is this vector a covariate?
}
