% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\docType{methods}
\name{is_exposure}
\alias{is_exposure}
\alias{is_exposure,context-method}
\alias{is_exposure,context,context-method}
\alias{is_exposure,described-method}
\alias{is_exposure,described,context-method}
\title{Is this vector an exposure?}
\usage{
is_exposure(x)

\S4method{is_exposure}{context}(x)

\S4method{is_exposure}{described}(x)
}
\arguments{
\item{x}{what to check}
}
\description{
Is this vector an exposure?
}
