% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\docType{methods}
\name{is_outcome}
\alias{is_outcome}
\alias{is_outcome,context-method}
\alias{is_outcome,context,context-method}
\alias{is_outcome,described-method}
\alias{is_outcome,described,context-method}
\title{Is this vector an outcome?}
\usage{
is_outcome(x)

\S4method{is_outcome}{context}(x)

\S4method{is_outcome}{described}(x)
}
\description{
Is this vector an outcome?
}
