% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\docType{methods}
\name{is_purpose}
\alias{is_purpose}
\alias{is_purpose,context-method}
\alias{is_purpose,context,context-method}
\alias{is_purpose,NULL-method}
\alias{is_purpose,NULL,context-method}
\title{Check a context}
\usage{
is_purpose(context, x)

\S4method{is_purpose}{context}(context, x)

\S4method{is_purpose}{`NULL`}(context, x)
}
\arguments{
\item{context}{a context}

\item{x}{what to check}
}
\description{
Check a context
}
