% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{is_stype}
\alias{is_stype}
\title{Check that an object is a stype vector}
\usage{
is_stype(object)
}
\arguments{
\item{object}{any \code{R} object}
}
\description{
Check that an object is a stype vector
}
