% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make_maybe_get_summary}
\alias{make_maybe_get_summary}
\title{Creates a function for getting a summary value}
\usage{
make_maybe_get_summary(summary_name, fn)
}
\arguments{
\item{summary_name}{name of the value in the stype's \code{data_summary}}

\item{fn}{the function to use in the case that the summary is not already computed.}
}
\description{
Creates a function for getting a summary value
}
\seealso{
maybe_get_data_summary_math
}
\keyword{internal}
