% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make_stype_constructor}
\alias{make_stype_constructor}
\title{Create a constructor function for stype vectors}
\usage{
make_stype_constructor(
  typeFUN,
  ptypeFUN,
  castFUN = ptypeFUN,
  dataFUN = identity
)
}
\arguments{
\item{typeFUN}{the \code{new_} function that creates vdctors of the type}

\item{ptypeFUN}{the prototype of the vector}

\item{castFUN}{optionally cast the data by \code{\link[vctrs]{vec_cast}}}

\item{dataFUN}{a function to apply to the data before \code{describe}ing the
the data. Defaults to \code{identity}.}
}
\description{
Create a constructor function for stype vectors
}
\keyword{internal}
