% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make_stype_ptype2}
\alias{make_stype_ptype2}
\title{Create a stype<->stype ptype2 function}
\usage{
make_stype_ptype2(constructorFUN)
}
\arguments{
\item{constructorFUN}{the constructor function to use.}
}
\description{
Create a stype<->stype ptype2 function
}
\keyword{internal}
