% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make_stype_restorator}
\alias{make_stype_restorator}
\title{Create a vec_restore function for basic stype types}
\usage{
make_stype_restorator(typeFUN)
}
\arguments{
\item{typeFUN}{the \code{new_} function that creates vdctors of the type}
}
\description{
Create a vec_restore function for basic stype types
}
\keyword{internal}
