% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{maybe_get_data_summary_math}
\alias{maybe_get_data_summary_math}
\title{Gets a summary value already computed in a data_summary}
\usage{
maybe_get_data_summary_math(summary_name, .fn, .x, ...)
}
\arguments{
\item{summary_name}{name of the value in the stype's \code{data_summary}}

\item{.fn}{A mathematical function from the base package, as a string.}

\item{.x}{A vector.}

\item{...}{Additional arguments passed to \code{.fn}.}
}
\description{
Gets a value from a stype vector's data_summary in the case the summary is
auto computed AND the \code{na.rm} argument is TRUE.
}
\keyword{internal}
