% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purpose.R, R/context.R
\name{modify_purpose}
\alias{modify_purpose}
\alias{add_study_role}
\alias{add_study_role,purpose,character-method}
\alias{add_study_role,purpose,purpose-method}
\alias{add_tags}
\alias{add_tags,purpose,character-method}
\alias{add_tags,purpose,purpose-method}
\alias{remove_study_role}
\alias{remove_study_role,purpose,character-method}
\alias{remove_study_role,purpose,purpose-method}
\alias{remove_tags}
\alias{remove_tags,purpose,character-method}
\alias{remove_tags,purpose,purpose-method}
\alias{add_study_role,context,character-method}
\alias{add_study_role,context,context-method}
\alias{add_study_role,stype,character-method}
\alias{add_study_role,stype,stype-method}
\alias{remove_study_role,context,character-method}
\alias{remove_study_role,context,context-method}
\alias{remove_study_role,stype,character-method}
\alias{remove_study_role,stype,stype-method}
\alias{add_tags,context,character-method}
\alias{add_tags,context,context-method}
\alias{add_tags,stype,character-method}
\alias{add_tags,stype,stype-method}
\alias{remove_tags,context,character-method}
\alias{remove_tags,context,context-method}
\alias{remove_tags,stype,character-method}
\alias{remove_tags,stype,stype-method}
\title{Modify a purpose}
\usage{
add_study_role(object, role)

\S4method{add_study_role}{purpose,character}(object, role)

add_tags(object, tags)

\S4method{add_tags}{purpose,character}(object, tags)

remove_study_role(object, role)

\S4method{remove_study_role}{purpose,character}(object, role)

remove_tags(object, tags)

\S4method{remove_tags}{purpose,character}(object, tags)

\S4method{add_study_role}{context,character}(object, role)

\S4method{add_study_role}{stype,character}(object, role)

\S4method{remove_study_role}{context,character}(object, role)

\S4method{remove_study_role}{stype,character}(object, role)

\S4method{add_tags}{context,character}(object, tags)

\S4method{add_tags}{stype,character}(object, tags)

\S4method{remove_tags}{context,character}(object, tags)

\S4method{remove_tags}{stype,character}(object, tags)
}
\arguments{
\item{object}{an object}

\item{role}{the roles to remove}

\item{tags}{the tags to add/remove}
}
\description{
Modify a purpose

Add a study role

Add tags

Remove a study role

Remove tags
}
