% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_nominal.R
\name{new_nominal}
\alias{new_nominal}
\alias{vec_type2.v_nominal}
\alias{vec_cast.v_nominal}
\title{v_nominal/ordered v_nominal S3 class}
\usage{
new_nominal(x = integer(), .levels = character(),
  .desc = description(), .context = context())

\method{vec_type2}{v_nominal}(x, y, ...)

\method{vec_cast}{v_nominal}(x, to, ...)
}
\arguments{
\item{x}{Integer values which index in to `levels`.}

\item{.levels}{Character vector of labels.}

\item{.desc}{a description}

\item{.context}{a context}
}
\description{
A [v_nominal] is an integer with attribute `levels`, a character vector. There
should be one level for each integer between 1 and `max(x)`.
An [ordered] v_nominal has the same properties as a v_nominal, but possesses
an extra class that marks levels as having a total ordering.
}
\details{
These functions help the base v_nominal and ordered v_nominal classes fit in to
the vctrs type system by providing constructors, coercion functions,
and casting functions. `new_v_nominal()` and `new_ordered()` are low-level
constructors - they only check that types, but not values, are valid, so
are for expert use only.
}
