% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{new_stype_vctr}
\alias{new_stype_vctr}
\title{The internal builder of (basic) stype vctrs}
\usage{
new_stype_vctr(
  x,
  .internal_name,
  .data_summary,
  .context,
  .extra_descriptors,
  .auto_compute_summary,
  .class,
  new_fun = vctrs::new_vctr,
  ...
)
}
\arguments{
\item{x}{the vector data}

\item{.internal_name}{\code{internal_name} attribute}

\item{.data_summary}{\code{\link{data_summary}} attribute}

\item{.context}{\code{\link{context}} attribute}

\item{.class}{passed to the \code{class} argument of (e.g.) the \code{new_fun}}

\item{new_fun}{the \code{vctrs} \verb{new_*} function to use. Defaults to
\code{\link[vctrs:new_vctr]{vctrs::new_vctr()}}. The first argument must the "data" argument called
".data" in new_vctr and "x" in new_factor and "fields" in new_rcrd.}

\item{.extra_descriptor}{\code{extra_descriptors} attribute a \code{list} of descriptor
functions}
}
\description{
Used instead of \code{vctrs::new_vctr} in each builder in order to keep the
attributes consistent across stype vctrs.
}
\keyword{internal}
