% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_rcensored.R
\name{pmin_v_rcensored}
\alias{pmin_v_rcensored}
\title{Combine Right-Censored Events}
\usage{
pmin_v_rcensored(
  ...,
  new_end_time = "strict",
  internal_name = "",
  context = methods::new("context"),
  extra_descriptors = list()
)
}
\arguments{
\item{...}{passed to other methods such as \code{as.character}}

\item{new_end_time}{either one of \code{"strict"} or \code{"min"}, or a postive scalar
value. In the case of \code{"strict"} the inputs are assumed to all have the
same \code{end_time} value (otherwise and error is thrown), and the \code{end_time}
of the output object is the same as that of the inputs. In the case of
\code{"min"}, the minimum \code{end_time} value among all of the inputs is the taken
to be the \code{end_time} of the output object. In the case of a positive scalar
input for \code{new_end_time}, the \code{end_time} of the output object is taken to
be the value of \code{new_end_time}. Note that an error is thrown if the value
of \code{new_end_time} is larger than the smallest \code{end_time} value among the
inputs since this situation could result in faulty data.}

\item{internal_name}{the internal name of the variable}

\item{context}{a \code{\link{context}}}

\item{extra_descriptors}{A \code{list} of \code{\link{descriptors}} functions
appended to the default \code{\link{descriptors}}.}
}
\value{
A \code{v_rcensored} object.
}
\description{
Create a right-censored event based on the minimum time to event among a
collection of right-censored events. In the event of ties for censoring or
outcome events, earlier inputs are given precedence over later inputs.
}
