% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lenses.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{view}
\alias{set}
\alias{over_map}
\alias{over}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{lenses}{\code{\link[lenses]{over}}, \code{\link[lenses]{over_map}}, \code{\link[lenses]{over_map}}, \code{\link[lenses]{set}}, \code{\link[lenses]{view}}}
}}

