% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableClass.R, R/aREADME.R
\docType{methods}
\name{show,variable-method}
\alias{show,variable-method}
\alias{show,variable,variable-method}
\alias{show-methods}
\title{Display an object}
\usage{
\S4method{show}{variable}(object)
}
\arguments{
\item{object}{the object to display}
}
\description{
Display an object
}
