% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aREADME.R
\docType{package}
\name{stype-package}
\alias{stype}
\alias{stype-package}
\title{Statistical algebraic data types}
\description{
stype (pronounced "stipe") uses the \link[vctrs]{vctrs} package
to implement algebraic data types that:
\itemize{
\item provide S3 classes aligned with statistical language: binary,
continuous, count, non-negative continuous, time to event, etc;
\item create variable-level metadata useful for sharing information
about a variable across applications;
\item add automatically generated summary statistics to the variable metadata.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/novisci/stype}
  \item Report bugs at \url{https://github.com/novisci/stype/issues}
}

}
\author{
\strong{Maintainer}: Bradley Saul \email{bsaul@novisci.com}

Authors:
\itemize{
  \item David Pritchard \email{dpritchard@novisci.com}
}

Other contributors:
\itemize{
  \item Catie Wiener \email{cwiener@novisci.com} [contributor]
  \item Ying Yu \email{yingyu@novisci.com} [contributor]
  \item Matt Phelan \email{mphelan@novisci.com} [contributor]
  \item Bassim Eledath \email{beledath@novisci.com} [contributor]
  \item Brendan Brown \email{bbrown@targetrwe.com} [contributor]
}

}
