% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{stype_abort}
\alias{stype_abort}
\title{Create custom \code{rlang::abort} Class for Stype Objects}
\usage{
stype_abort(msg = NULL, class = NULL, ...)
}
\arguments{
\item{msg}{a string containing the error message.}

\item{class}{a character vector with error classes, if any, to add in
addition to "stype_err".}

\item{...}{additional arguments passed to \code{abort}}
}
\description{
modeled after \code{vctrs:::stop_vctrs}
}
\keyword{internal}
