% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\docType{class}
\name{stype_classes}
\alias{stype_classes}
\alias{groupVar-class}
\alias{stypeWeightVar-class}
\alias{weightVar-class}
\alias{categorical-class}
\alias{Missing-class}
\alias{maybeGroup-class}
\alias{maybeWeight-class}
\alias{maybeDescriptor-class}
\alias{describable-class}
\alias{simpleStype-class}
\alias{complexStype-class}
\alias{v_stype-class}
\alias{tbl_stype-class}
\alias{stype-class}
\title{Stype classes}
\description{
S4 classes uses in defining methods.
}
\keyword{internal}
