% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lenses.R
\docType{data}
\name{stype_lenses}
\alias{stype_lenses}
\alias{stype_df_l}
\alias{binary_l}
\alias{count_l}
\alias{rcensored_l}
\alias{outcome_l}
\alias{covariate_l}
\alias{weight_l}
\alias{constant_l}
\alias{not_constant_l}
\alias{data_summmary_l}
\alias{vec_data_l}
\alias{context_l}
\alias{internal_name_l}
\title{Lenses for stype vectors}
\format{An object of class \code{lens} of length 2.}
\usage{
stype_df_l(predicate)

binary_l

count_l

rcensored_l

outcome_l

covariate_l

weight_l

constant_l

not_constant_l

data_summmary_l

vec_data_l

context_l

internal_name_l
}
\arguments{
\item{predicate}{a predicate function to apply to each element of the list 
(or each column of a \code{data.frame})}
}
\description{
Lenses for stype vectors
}
\section{Functions}{
\itemize{
\item \code{stype_df_l}: Creates a lens for stype variables with a list-like 
structure

\item \code{binary_l}: view/set \code{\link{v_binary}} stype vectors

\item \code{count_l}: view/set \code{\link{v_count}} stype vectors

\item \code{rcensored_l}: view/set \code{\link{v_rcensored}} stype vectors

\item \code{outcome_l}: view/set \code{outcome} stype vectors

\item \code{covariate_l}: view/set \code{covariate} stype vectors

\item \code{weight_l}: view/set \code{weight} stype vectors

\item \code{constant_l}: view/set constant stype vectors

\item \code{not_constant_l}: view/set not constant stype vectors

\item \code{data_summmary_l}: view/set the \code{data_summary} object of a stype vector

\item \code{vec_data_l}: view/set the \code{\link[vctrs]{vec_data}} of a
\code{stype}. Casts to the original type when \code{set}.

\item \code{context_l}: view/set the \code{\link{context}} of a \code{stype}.

\item \code{internal_name_l}: view/set the \code{internal_name} of a \code{stype}.
}}

\keyword{datasets}
