% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{stype_str}
\alias{stype_str}
\title{String Representation of the Statistical Type}
\usage{
stype_str(x)
}
\arguments{
\item{x}{A stype object.}
}
\value{
A string.
}
\description{
Provides a short phrase identifying the statistical type.
}
