% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{swap_function}
\alias{swap_function}
\title{Swap the called function for a different function}
\usage{
swap_function(cl, fn)
}
\arguments{
\item{cl}{a call}

\item{fn}{the new function}
}
\description{
Swap the called function for a different function
}
\keyword{internal}
