% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_analysis.R
\name{tbl_analysis}
\alias{tbl_analysis}
\alias{analysis}
\alias{is_analysis}
\alias{as_canonical.tbl_analysis}
\title{Analysis}
\usage{
analysis(
  x = tibble::tibble(),
  internal_name = "",
  context,
  extra_descriptors = list(),
  modifiers = list(function(dt) identity(dt))
)

is_analysis(x)

\method{as_canonical}{tbl_analysis}(x)
}
\arguments{
\item{x}{a \code{tibble}}

\item{internal_name}{the internal name of the variable}

\item{context}{a \code{\link{context}}}

\item{extra_descriptors}{A \code{list} of \code{\link{descriptors}} functions
appended to the default \code{\link{descriptors}}.}

\item{modifiers}{a \code{list} of functions sequentially modify \code{x}.
Defaults to simply the \code{identity} function}
}
\description{
{
Prototype: \code{\link{tibble}}
}
}
\seealso{
Other stype types: 
\code{\link{v_binary}},
\code{\link{v_continuous_nonneg}},
\code{\link{v_continuous}},
\code{\link{v_count}},
\code{\link{v_nominal}},
\code{\link{v_ordered}},
\code{\link{v_proportion}},
\code{\link{v_rcensored}}
}
\concept{stype types}
