% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_binary.R
\name{v_binary}
\alias{v_binary}
\alias{is_binary}
\alias{as_binary}
\alias{as_canonical.v_binary}
\title{Binary vectors}
\usage{
v_binary(x = logical(), internal_name = "", context)

is_binary(x)

as_binary(x)

\method{as_canonical}{v_binary}(x)
}
\arguments{
\item{x}{a \code{logical} vector}

\item{internal_name}{the internal name of the variable}

\item{context}{a \code{\link{context}}}
}
\description{
Some desc

constructor function for binary objects
}
