% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_character.R
\name{v_character}
\alias{v_character}
\alias{is_character}
\alias{as_character}
\alias{as_canonical.v_character}
\title{Character vectors}
\usage{
v_character(x = character(), internal_name = "", context)

is_character(x)

as_character(x)

\method{as_canonical}{v_character}(x)
}
\arguments{
\item{x}{a \code{character} vector}

\item{internal_name}{the internal name of the variable}

\item{context}{a \code{\link{context}}}
}
\description{
Some desc

constructor function for character objects
}
