% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_continuous.R
\name{v_continuous}
\alias{v_continuous}
\alias{cont}
\alias{is_continuous}
\alias{as_continuous}
\alias{as.character.v_continuous}
\alias{as_canonical.v_continuous}
\title{Continuous vectors}
\usage{
v_continuous(
  x = ptypeFUN(),
  internal_name = "",
  context,
  extra_descriptors = list()
)

cont(x = ptypeFUN(), internal_name = "", context, extra_descriptors = list())

is_continuous(x)

as_continuous(x)

\method{as.character}{v_continuous}(x, ...)

\method{as_canonical}{v_continuous}(x)
}
\arguments{
\item{x}{a \code{double} vector}

\item{internal_name}{the internal name of the variable}

\item{context}{a \code{\link{context}}}

\item{extra_descriptors}{A \code{list} of \code{\link{descriptors}} functions
appended to the default \code{\link{descriptors}}.}

\item{...}{passed to other methods such as \code{as.character}}
}
\description{
{
Support: \eqn{\mathbf{R}}{the Reals}* (plus \code{\link{NA_real_}})

Prototype: \code{\link{double}}
\itemize{
\item 
\itemize{
\item i.e. floating-point number
}
}
}
}
\seealso{
Other stype types: 
\code{\link{tbl_analysis}},
\code{\link{v_binary}},
\code{\link{v_character}},
\code{\link{v_continuous_nonneg}},
\code{\link{v_count}},
\code{\link{v_event_time}},
\code{\link{v_nominal}},
\code{\link{v_ordered}},
\code{\link{v_rcensored}}
}
\concept{stype types}
