% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_continuous_nonneg.R
\name{v_continuous_nonneg}
\alias{v_continuous_nonneg}
\alias{new_continuous_nonneg}
\alias{is_continuous_nonneg}
\alias{is_nonneg}
\alias{as_nonneg_continuous}
\title{continuous_nonneg vectors}
\usage{
new_continuous_nonneg(x = double(), .desc = description(),
  .context = context())

v_continuous_nonneg(x = double(), context)

is_continuous_nonneg(x)

is_nonneg(x)

as_nonneg_continuous(x)
}
\arguments{
\item{x}{nonnegative \code{double}}

\item{.desc}{a \\code{\link{description}}}

\item{.context}{a \code{\link{context}}}

\item{context}{a \code{\link{context}}}
}
\description{
Some desc

constructor function for count objects
}
