% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_nominal.R
\name{v_nominal}
\alias{v_nominal}
\alias{new_nominal}
\alias{is_nominal}
\title{v_nominal S3 class}
\usage{
new_nominal(x = integer(), .levels = character(),
  .internal_name = character(), .data_summary = data_summary(),
  .context = context())

v_nominal(x = factor(), internal_name = "", context)

is_nominal(x)
}
\arguments{
\item{x}{a \code{factor}}

\item{.levels}{Character vector of labels.}

\item{.context}{a \code{\link{context}}}

\item{context}{a \code{\link{context}}}
}
\description{
A nominal (categorial) variable
}
