% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_ordered.R
\name{v_ordered}
\alias{v_ordered}
\alias{new_ordered}
\alias{is_ordered}
\title{v_ordered S3 class}
\usage{
new_ordered(x = integer(), .levels = character(),
  .desc = description(), .context = context())

v_ordered(x = factor(ordered = TRUE), context)

is_ordered(x)
}
\arguments{
\item{x}{a \code{factor}}

\item{.levels}{Character vector of labels.}

\item{.desc}{a \\code{\link{description}}}

\item{.context}{a \code{\link{context}}}

\item{context}{a \code{\link{context}}}
}
\description{
An ordered (categorial) variable
}
