% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_proportion.R
\name{v_proportion}
\alias{v_proportion}
\alias{prop}
\alias{is_proportion}
\alias{is_prop}
\alias{as_proportion}
\alias{as_canonical.v_proportion}
\title{Proportion vectors}
\usage{
v_proportion(
  x = ptypeFUN(),
  internal_name = "",
  context,
  auto_compute_summary = auto_compute_default,
  extra_descriptors = list()
)

prop(
  x = ptypeFUN(),
  internal_name = "",
  context,
  auto_compute_summary = auto_compute_default,
  extra_descriptors = list()
)

is_proportion(x)

is_prop(x)

as_proportion(x)

\method{as_canonical}{v_proportion}(x)
}
\arguments{
\item{x}{nonnegative \code{double}}

\item{internal_name}{the internal name of the variable}

\item{context}{a \code{\link{context}}}

\item{auto_compute_summary}{an indicator of whether the \code{data_summary} is
automatically computed whenever a vector is initialized, subset, or
concatenated. Defaults to \code{TRUE}. If this option is set to \code{FALSE},
then \code{\link{get_data_summary}} is the only way to compute the summary.
The \code{\link{data_summary_l}} lens will return an empty \code{data_summary}.}

\item{extra_descriptors}{A \code{list} of \code{\link{descriptors}} functions
appended to the default \code{\link{descriptors}}.}
}
\description{
Constructors and methods for a proportion  data type.
\code{v_proportion} and \code{prop} are synonyms that each create a
new \code{v_proportion} object subclassed from \code{v_continuous} and
\code{vctrs_vctr}. \cr\cr
Support: \eqn{\mathbf{[0, 1]}}* (plus \code{\link{NA_real_}}) \cr
Prototype: \code{\link{double}} \cr
\emph{}* - i.e. floating-point number
}
\examples{
# Example data
src_dbl <- c(1, 0.5, .01, 0, NA_real_)

# Constructor for the `v_proportion` class. One can also use `prop`
# which is a synonym for the `v_proportion` function.
v <- v_proportion(
  x = src_dbl,
  internal_name = "v_example",
  context = context(
    short_label = "important_var",
    long_label  = "Very important variable"
  ),
  extra_descriptors = list()
)

# Helper functions and methods
is_proportion(v)
as_proportion(src_dbl)
as.character(v)
as_canonical(v)
}
\seealso{
Other stype types: 
\code{\link{tbl_analysis}},
\code{\link{v_binary}},
\code{\link{v_continuous_nonneg}},
\code{\link{v_continuous}},
\code{\link{v_count}},
\code{\link{v_nominal}},
\code{\link{v_ordered}},
\code{\link{v_rcensored}}
}
\concept{stype types}
