% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_rcensored.R
\name{v_rcensored_accessors}
\alias{v_rcensored_accessors}
\alias{get_time}
\alias{get_censored}
\alias{get_censor_reason}
\alias{get_outcome}
\alias{get_outcome_reason}
\title{Get elements from v_rcensored}
\usage{
get_time(x)

get_censored(x)

get_censor_reason(x)

get_outcome(x)

get_outcome_reason(x)
}
\arguments{
\item{x}{a \code{\link{v_rcensored}} vector}
}
\description{
Get elements from v_rcensored
}
