% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableClass.R
\docType{class}
\name{variable-class}
\alias{variable-class}
\alias{variable}
\title{The variable R object}
\description{
Extends R's \code{\link{vector-class}} to provide objects with additional information. 
The vector object must be one of the following basic classes: \code{"numeric", "double", 
"integer", "logical", "character",}, or \code{\link{factor}}, or \code{\link{Date}}.
 A variable object can used as if it was a vector.
}
\section{Slots}{

\describe{
\item{\code{name}}{machine readable name}

\item{\code{short_label}}{a short label}

\item{\code{long_label}}{a longer label}

\item{\code{description}}{a \linkS4class{description} object containing summary statistics}
}}

