% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_count.R, R/v_continuous.R, R/v_binary.R
\name{vec_arith}
\alias{vec_arith}
\alias{vec_arith.v_count}
\alias{vec_arith.integer}
\alias{vec_arith.v_continuous}
\alias{vec_arith.double}
\alias{vec_arith.v_binary}
\title{Arithmetic ops}
\usage{
\method{vec_arith}{v_count}(op, x, y)

\method{vec_arith}{integer}(op, x, y)

\method{vec_arith}{v_continuous}(op, x, y)

\method{vec_arith}{double}(op, x, y)

\method{vec_arith}{v_binary}(op, x, y)
}
\arguments{
\item{op}{An arithmetic operator as a string}

\item{x}{A pair of vectors. For \code{!}, unary \code{+} and unary \code{-}, \code{y} will be
a sentinel object of class \code{MISSING}, as created by \code{MISSING()}.}

\item{y}{A pair of vectors. For \code{!}, unary \code{+} and unary \code{-}, \code{y} will be
a sentinel object of class \code{MISSING}, as created by \code{MISSING()}.}
}
\description{
Arithmetic ops
}
