% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{weight}
\alias{weight}
\alias{weight,stype,stypeWeightVar,maybeDescriptor-method}
\alias{weight,stype,weightVar,maybeDescriptor-method}
\title{Change Weights for stype Vector Elements}
\usage{
weight(x, w, .descriptors = list())

\S4method{weight}{stype,stypeWeightVar,maybeDescriptor}(x, w, .descriptors = list())

\S4method{weight}{stype,weightVar,maybeDescriptor}(x, w, .descriptors = list())
}
\arguments{
\item{x}{a vector a values}

\item{w}{a vector of weights (optional)}

\item{.descriptors}{an optional list of functions. Each function should have
a signature of the form \verb{function(v, ...)}, and where \code{v} is expected to be
of the result of calling \code{as_canonical(x)} on the \code{x} input to \code{describe},
and \code{...} is expected to be the unchanged value of the input to \code{...} in
\code{describe}. If the input list is named then elements in the return object
corresponding to the result of evaluating the supplied functions will have
the corresponding names.}
}
\description{
Return a stype vector where the elements are weighted according to the
specified inputs. By default, all of the weights for a stype vector are
effectively constant.
}
