% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_rcensored.R
\name{as_Surv}
\alias{as_Surv}
\title{Cast a v_rcensored type to a Surv object}
\usage{
as_Surv(x, censor_as_event = FALSE, multiple_endpoint = FALSE)
}
\arguments{
\item{x}{a \code{\link{v_rcensored}} object}

\item{censor_as_event}{an indicator to treat censoring as the event of
interest (such as when estimating censoring probabilities). When
\code{TRUE}, all outcomes are treated as a single censoring event and
censoring events as a single outcome. Defaults to \code{FALSE}.}

\item{multiple_endpoint}{an indicator to treat the outcomes as multiple
endpoint data, otherwise all outcomes are treated as being the same event
type. Defaults to \code{FALSE}.}
}
\description{
Supports (hopefully obviously) \code{type = "right"} in \code{survival::Surv}
}
