% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_rcensored.R
\name{as_Surv}
\alias{as_Surv}
\title{Cast a v_rcensored type to a Surv object}
\usage{
as_Surv(x)
}
\arguments{
\item{x}{a \code{\link{v_rcensored}} object}
}
\description{
Supports (hopefully obviously) \code{type = "right"} in \code{survival::Surv}
}
