% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\docType{class}
\name{context}
\alias{context}
\title{A context contains study design specific information about a variable}
\description{
A context contains study design specific information about a variable
}
\section{Slots}{

\describe{
\item{\code{name}}{a name}

\item{\code{short_label}}{a short_label}

\item{\code{long_label}}{a long_label}

\item{\code{purpose}}{a purpose}
}}

