% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{get_context}
\alias{get_context}
\alias{get_context,stype-method}
\alias{get_short_label}
\alias{get_short_label,context-method}
\alias{get_short_label,context,context-method}
\alias{get_short_label,stype-method}
\alias{get_short_label,stype,stype-method}
\alias{get_long_label}
\alias{get_long_label,context-method}
\alias{get_long_label,context,context-method}
\alias{get_long_label,stype-method}
\alias{get_long_label,stype,stype-method}
\alias{get_description}
\alias{get_description,context-method}
\alias{get_description,context,context-method}
\alias{get_description,stype-method}
\alias{get_description,stype,stype-method}
\alias{get_derivation}
\alias{get_derivation,context-method}
\alias{get_derivation,context,context-method}
\alias{get_derivation,stype-method}
\alias{get_derivation,stype,stype-method}
\alias{get_purpose}
\alias{get_purpose,context-method}
\alias{get_purpose,context,context-method}
\alias{get_purpose,stype-method}
\alias{get_purpose,stype,stype-method}
\alias{get_security_type}
\alias{get_security_type,context-method}
\alias{get_security_type,context,context-method}
\alias{get_security_type,stype-method}
\alias{get_security_type,stype,stype-method}
\alias{set_context}
\alias{set_context,stype,context-method}
\alias{set_context,stype,stype-method}
\alias{set_short_label}
\alias{set_short_label,context,character-method}
\alias{set_short_label,context,context-method}
\alias{set_short_label,stype,character-method}
\alias{set_short_label,stype,stype-method}
\alias{set_long_label}
\alias{set_long_label,context,character-method}
\alias{set_long_label,context,context-method}
\alias{set_long_label,stype,character-method}
\alias{set_long_label,stype,stype-method}
\alias{set_description}
\alias{set_description,context,character-method}
\alias{set_description,context,context-method}
\alias{set_description,stype,character-method}
\alias{set_description,stype,stype-method}
\alias{set_derivation}
\alias{set_derivation,context,character-method}
\alias{set_derivation,context,context-method}
\alias{set_derivation,stype,character-method}
\alias{set_derivation,stype,stype-method}
\alias{set_purpose}
\alias{set_purpose,context,purpose-method}
\alias{set_purpose,context,context-method}
\alias{set_purpose,stype,purpose-method}
\alias{set_purpose,stype,stype-method}
\alias{set_security_type}
\alias{set_security_type,context,character-method}
\alias{set_security_type,context,context-method}
\alias{set_security_type,stype,character-method}
\alias{set_security_type,stype,stype-method}
\title{Get an object's context or elements thereof}
\usage{
get_context(x)

\S4method{get_context}{stype}(x)

get_short_label(x)

\S4method{get_short_label}{context}(x)

\S4method{get_short_label}{stype}(x)

get_long_label(x)

\S4method{get_long_label}{context}(x)

\S4method{get_long_label}{stype}(x)

get_description(x)

\S4method{get_description}{context}(x)

\S4method{get_description}{stype}(x)

get_derivation(x)

\S4method{get_derivation}{context}(x)

\S4method{get_derivation}{stype}(x)

get_purpose(x)

\S4method{get_purpose}{context}(x)

\S4method{get_purpose}{stype}(x)

get_security_type(x)

\S4method{get_security_type}{context}(x)

\S4method{get_security_type}{stype}(x)

\S4method{get_purpose}{stype}(x)

set_context(x, to)

\S4method{set_context}{stype,context}(x, to)

set_short_label(x, to)

\S4method{set_short_label}{context,character}(x, to)

\S4method{set_short_label}{stype,character}(x, to)

set_long_label(x, to)

\S4method{set_long_label}{context,character}(x, to)

\S4method{set_long_label}{stype,character}(x, to)

set_description(x, to)

\S4method{set_description}{context,character}(x, to)

\S4method{set_description}{stype,character}(x, to)

set_derivation(x, to)

\S4method{set_derivation}{context,character}(x, to)

\S4method{set_derivation}{stype,character}(x, to)

set_purpose(x, to)

\S4method{set_purpose}{context,purpose}(x, to)

\S4method{set_purpose}{stype,purpose}(x, to)

set_security_type(x, to)

\S4method{set_security_type}{context,character}(x, to)

\S4method{set_security_type}{stype,character}(x, to)
}
\arguments{
\item{x}{a stype object}

\item{to}{what to set the element to}
}
\description{
Get an object's context or elements thereof
}
