% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\docType{class}
\name{data_summary-class}
\alias{data_summary-class}
\alias{data_summary}
\title{A call to describe returns a data_summary object}
\description{
A call to describe returns a data_summary object
}
