\name{DenverSugar-package}
\alias{DenverSugar-package}
\alias{DenverSugar}
\docType{package}
\title{
\packageTitle{DenverSugar}
}
\description{
\packageDescription{DenverSugar}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DenverSugar}
\packageIndices{DenverSugar}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{DenverSugar}

Maintainer: \packageMaintainer{DenverSugar}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
