% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\docType{methods}
\name{describe}
\alias{describe}
\alias{describe,variable,maybeGroup,maybeWeight-method}
\title{Describe a variable}
\usage{
describe(x, g = NULL, w = NULL, ...)

\S4method{describe}{variable,maybeGroup,maybeWeight}(x, g = NULL,
  w = NULL, ...)
}
\arguments{
\item{x}{a vector a values}

\item{g}{a vector a groupings (optional)}

\item{w}{a vector of weights (optional)}

\item{...}{additional arguments}
}
\description{
Describe a variable
}
