% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\docType{methods}
\name{describe}
\alias{describe}
\alias{describe,variable,maybeGroup,maybeWeight,maybeDescriptor-method}
\title{Describe a variable}
\usage{
describe(x, g = NULL, w = NULL, .descriptors, ...)

\S4method{describe}{variable,maybeGroup,maybeWeight,maybeDescriptor}(x,
  g = NULL, w = NULL, .descriptors, ...)
}
\arguments{
\item{x}{a vector a values}

\item{g}{a vector a groupings (optional)}

\item{w}{a vector of weights (optional)}

\item{.descriptors}{an (optional) list of lambda functions}

\item{...}{additional arguments}
}
\description{
Describe a variable
}
