% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\docType{data}
\name{descriptors}
\alias{descriptors}
\alias{standardDescriptors}
\alias{logicalDescriptors}
\alias{wlogicalDescriptors}
\alias{categoricalDescriptors}
\alias{characterDescriptors}
\alias{numericDescriptors}
\alias{wnumericDescriptors}
\alias{groupedDescriptors}
\alias{rcensoredDescriptors}
\alias{analysisDescriptors}
\title{Descriptors}
\format{
An object of class \code{list} of length 6.

An object of class \code{list} of length 4.

An object of class \code{list} of length 1.

An object of class \code{list} of length 3.

An object of class \code{list} of length 3.

An object of class \code{list} of length 8.

An object of class \code{list} of length 1.

An object of class \code{list} of length 1.

An object of class \code{list} of length 4.

An object of class \code{list} of length 4.
}
\usage{
standardDescriptors

logicalDescriptors

wlogicalDescriptors

categoricalDescriptors

characterDescriptors

numericDescriptors

wnumericDescriptors

groupedDescriptors

rcensoredDescriptors

analysisDescriptors
}
\description{
Lists of functions to used to \code{\link{describe}} a \code{describeable}
vector to create a \code{\link{data_summary}} object. A \code{descriptor} is
a function of \code{x} (and optionally \code{g} (a grouping vector and/or
\code{w} a numeric weight vector)).

\code{describable} vectors include: integer, logical, numeric, factor,
ordered, and v_rcensored.
}
\keyword{datasets}
\keyword{internal}
