% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{get_purpose}
\alias{get_purpose}
\alias{get_purpose,context-method}
\alias{get_purpose,context,context-method}
\alias{get_purpose,stype-method}
\alias{get_purpose,stype,stype-method}
\title{Get an object's purpose}
\usage{
get_purpose(x)

\S4method{get_purpose}{context}(x)

\S4method{get_purpose}{stype}(x)
}
\arguments{
\item{x}{a stype object}
}
\description{
Get an object's purpose
}
