% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\docType{methods}
\name{getDescriptors}
\alias{getDescriptors}
\alias{getDescriptors,variable-method}
\alias{getDescriptors,logical-method}
\alias{getDescriptors,character-method}
\alias{getDescriptors,numeric-method}
\alias{getDescriptors,NULL-method}
\title{Descriptor}
\usage{
getDescriptors(x)

\S4method{getDescriptors}{variable}(x)

\S4method{getDescriptors}{logical}(x)

\S4method{getDescriptors}{character}(x)

\S4method{getDescriptors}{numeric}(x)

\S4method{getDescriptors}{`NULL`}(x)
}
\arguments{
\item{x}{a \linkS4class{variable} or vector}
}
\description{
A function that returns a list of functions to be applied to a variable
}
