% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purpose.R, R/context.R
\docType{methods}
\name{is_study_role}
\alias{is_study_role}
\alias{is_study_role,purpose-method}
\alias{is_study_role,purpose,purpose-method}
\alias{is_study_role,NULL-method}
\alias{is_study_role,NULL,NULL-method}
\alias{is_identifier}
\alias{is_identifier,purpose-method}
\alias{is_identifer,purpose,purpose-method}
\alias{is_index}
\alias{is_index,purpose-method}
\alias{is_index,purpose,purpose-method}
\alias{is_outcome}
\alias{is_outcome,purpose-method}
\alias{is_outcome,purpose,purpose-method}
\alias{is_censoring}
\alias{is_censoring,purpose-method}
\alias{is_censoring,purpose,purpose-method}
\alias{is_competing}
\alias{is_competing,purpose-method}
\alias{is_competing,purpose,purpose-method}
\alias{is_exposure}
\alias{is_exposure,purpose-method}
\alias{is_exposure,purpose,purpose-method}
\alias{is_covariate}
\alias{is_covariate,purpose-method}
\alias{is_covariate,purpose,purpose-method}
\alias{is_weight}
\alias{is_weight,purpose-method}
\alias{is_other,purpose,purpose-method}
\alias{is_intermediate}
\alias{is_intermediate,purpose-method}
\alias{is_other}
\alias{is_other,purpose-method}
\alias{is_study_role,context-method}
\alias{is_study_role,context,context-method}
\alias{is_study_role,described-method}
\alias{is_study_role,described,described-method}
\alias{is_identifier,context-method}
\alias{is_identifier,context,context-method}
\alias{is_identifier,described-method}
\alias{is_identifier,described,described-method}
\alias{is_index,context-method}
\alias{is_index,context,context-method}
\alias{is_index,described-method}
\alias{is_index,described,described-method}
\alias{is_outcome,context-method}
\alias{is_outcome,context,context-method}
\alias{is_outcome,described-method}
\alias{is_outcome,described,described-method}
\alias{is_censoring,context-method}
\alias{is_censoring,context,context-method}
\alias{is_censoring,described-method}
\alias{is_censoring,described,described-method}
\alias{is_competing,context-method}
\alias{is_competing,context,context-method}
\alias{is_competing,described-method}
\alias{is_competing,described,described-method}
\alias{is_exposure,context-method}
\alias{is_exposure,context,context-method}
\alias{is_exposure,described-method}
\alias{is_exposure,described,described-method}
\alias{is_covariate,context-method}
\alias{is_covariate,context,context-method}
\alias{is_covariate,described-method}
\alias{is_covariate,described,described-method}
\alias{is_weight,context-method}
\alias{is_weight,context,context-method}
\alias{is_weight,described-method}
\alias{is_weight,described,described-method}
\alias{is_intermediate,context-method}
\alias{is_intermediate,context,context-method}
\alias{is_intermediate,described-method}
\alias{is_intermediate,described,described-method}
\alias{is_other,context-method}
\alias{is_other,context,context-method}
\alias{is_other,described-method}
\alias{is_other,described,described-method}
\title{Check a study role}
\usage{
is_study_role(object, what)

\S4method{is_study_role}{purpose}(object, what)

\S4method{is_study_role}{`NULL`}(object, what)

is_identifier(object)

\S4method{is_identifier}{purpose}(object)

is_index(object)

\S4method{is_index}{purpose}(object)

is_outcome(object)

\S4method{is_outcome}{purpose}(object)

is_censoring(object)

\S4method{is_censoring}{purpose}(object)

is_competing(object)

\S4method{is_competing}{purpose}(object)

is_exposure(object)

\S4method{is_exposure}{purpose}(object)

is_covariate(object)

\S4method{is_covariate}{purpose}(object)

is_weight(object)

\S4method{is_weight}{purpose}(object)

is_intermediate(object)

\S4method{is_intermediate}{purpose}(object)

is_other(object)

\S4method{is_other}{purpose}(object)

\S4method{is_study_role}{context}(object, what)

\S4method{is_study_role}{described}(object, what)

\S4method{is_identifier}{context}(object)

\S4method{is_identifier}{described}(object)

\S4method{is_index}{context}(object)

\S4method{is_index}{described}(object)

\S4method{is_outcome}{context}(object)

\S4method{is_outcome}{described}(object)

\S4method{is_censoring}{context}(object)

\S4method{is_censoring}{described}(object)

\S4method{is_competing}{context}(object)

\S4method{is_competing}{described}(object)

\S4method{is_exposure}{context}(object)

\S4method{is_exposure}{described}(object)

\S4method{is_covariate}{context}(object)

\S4method{is_covariate}{described}(object)

\S4method{is_weight}{context}(object)

\S4method{is_weight}{described}(object)

\S4method{is_intermediate}{context}(object)

\S4method{is_intermediate}{described}(object)

\S4method{is_other}{context}(object)

\S4method{is_other}{described}(object)
}
\arguments{
\item{object}{an object}

\item{what}{what to role to look for}
}
\description{
Check a study role
}
