% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aREADME.R
\docType{package}
\name{stype-package}
\alias{stype}
\alias{stype-package}
\title{Statistical algebraic data types}
\description{
\code{stype} (pronounced "stipe") uses the \link[vctrs]{vctrs} package
to implement algebraic data types that:
}
\details{
\itemize{
\item provide S3 classes aligned with statistical language: binary,
continuous, count, non-negative continuous, time to event, etc;
\item create variable-level metadata useful for sharing information
about a variable across applications;
\item add automatically generated summary statistics to the variable metadata.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.novisci.com/nsStat/stype}
  \item Report bugs at \url{https://gitlab.novisci.com/nsStat/stype/issues}
}

}
