% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lenses.R
\docType{data}
\name{stype_lenses}
\alias{stype_lenses}
\alias{stype_df_l}
\alias{binary_l}
\alias{character_l}
\alias{continuous_l}
\alias{continuous_nonneg_l}
\alias{count_l}
\alias{event_time_l}
\alias{nominal_l}
\alias{ordered_l}
\alias{rcensored_l}
\alias{covariate_l}
\alias{censoring_l}
\alias{competing_l}
\alias{exposure_l}
\alias{identifier_l}
\alias{index_l}
\alias{intermediate_l}
\alias{other_l}
\alias{outcome_l}
\alias{weight_l}
\alias{constant_l}
\alias{not_constant_l}
\alias{tag_l}
\alias{data_summary_l}
\alias{vec_data_l}
\alias{context_l}
\alias{internal_name_l}
\alias{extra_descriptors_l}
\alias{derivation_l}
\alias{short_label_l}
\alias{long_label_l}
\alias{description_l}
\alias{security_type_l}
\alias{purpose_l}
\alias{study_role_l}
\alias{tags_l}
\title{Lenses for stype vectors}
\format{
An object of class \code{lens} of length 2.

An object of class \code{lens} of length 2.

An object of class \code{lens} of length 2.

An object of class \code{lens} of length 2.

An object of class \code{lens} of length 2.

An object of class \code{lens} of length 2.

An object of class \code{lens} of length 2.

An object of class \code{lens} of length 2.

An object of class \code{lens} of length 2.

An object of class \code{lens} of length 2.

An object of class \code{lens} of length 2.

An object of class \code{lens} of length 2.

An object of class \code{lens} of length 2.

An object of class \code{lens} of length 2.

An object of class \code{lens} of length 2.

An object of class \code{lens} of length 2.

An object of class \code{lens} of length 2.

An object of class \code{lens} of length 2.

An object of class \code{lens} of length 2.

An object of class \code{lens} of length 2.

An object of class \code{lens} of length 2.

An object of class \code{lens} of length 2.

An object of class \code{lens} of length 2.

An object of class \code{lens} of length 2.

An object of class \code{lens} of length 2.

An object of class \code{lens} of length 2.

An object of class \code{lens} of length 2.

An object of class \code{lens} of length 2.

An object of class \code{lens} of length 2.

An object of class \code{lens} of length 2.

An object of class \code{lens} of length 2.

An object of class \code{lens} of length 2.

An object of class \code{lens} of length 2.

An object of class \code{lens} of length 2.
}
\usage{
stype_df_l(predicate)

binary_l

character_l

continuous_l

continuous_nonneg_l

count_l

event_time_l

nominal_l

ordered_l

rcensored_l

covariate_l

censoring_l

competing_l

exposure_l

identifier_l

index_l

intermediate_l

other_l

outcome_l

weight_l

constant_l

not_constant_l

tag_l(tags)

data_summary_l

vec_data_l

context_l

internal_name_l

extra_descriptors_l

derivation_l

short_label_l

long_label_l

description_l

security_type_l

purpose_l

study_role_l

tags_l
}
\arguments{
\item{predicate}{a predicate function to apply to each element of the list
(or each column of a \code{data.frame})}

\item{tags}{a \code{character} vector of tags}
}
\description{
Lenses for stype vectors
}
\section{Functions}{
\itemize{
\item \code{stype_df_l}: Creates a lens for stype variables with a list-like
structure

\item \code{binary_l}: view/set \code{\link{v_binary}} stype vectors

\item \code{character_l}: view/set \code{\link{v_character}} stype vectors

\item \code{continuous_l}: view/set \code{\link{v_continuous}} stype vectors

\item \code{continuous_nonneg_l}: view/set \code{\link{v_continuous_nonneg}} stype vectors

\item \code{count_l}: view/set \code{\link{v_count}} stype vectors

\item \code{event_time_l}: view/set \code{\link{v_event_time}} stype vectors

\item \code{nominal_l}: view/set \code{\link{v_nominal}} stype vectors

\item \code{ordered_l}: view/set \code{\link{v_ordered}} stype vectors

\item \code{rcensored_l}: view/set \code{\link{v_rcensored}} stype vectors

\item \code{covariate_l}: view/set stype vectors with \code{covariate} role

\item \code{censoring_l}: view/set stype vectors with \code{censoring} role

\item \code{competing_l}: view/set stype vectors with \code{competing} role

\item \code{exposure_l}: view/set stype vectors with \code{exposure} role

\item \code{identifier_l}: view/set stype vectors with \code{identifier} role

\item \code{index_l}: view/set stype vectors with \code{index} role

\item \code{intermediate_l}: view/set stype vectors with \code{is_intermediate} role

\item \code{other_l}: view/set stype vectors with \code{other} role

\item \code{outcome_l}: view/set stype vectors with \code{outcome} role

\item \code{weight_l}: view/set stype vectors with  \code{weight} role

\item \code{constant_l}: view/set constant stype vectors

\item \code{not_constant_l}: view/set not constant stype vectors

\item \code{tag_l}: view/set stype vectors tagged with \code{tags}

\item \code{data_summary_l}: view/set the \code{data_summary} object of a stype vector

\item \code{vec_data_l}: view/set the \code{\link[vctrs]{vec_data}} of a
\code{stype}. Casts to the original type when \code{set}.

\item \code{context_l}: view/set the \code{\link{context}} of a \code{stype}.

\item \code{internal_name_l}: view/set the \code{internal_name} of a \code{stype}.

\item \code{extra_descriptors_l}: view/set the \code{extra_descriptors} of a \code{stype}.

\item \code{derivation_l}: view/set the \code{derivation} slot of a \code{stype}
\code{\link{context}}.

\item \code{short_label_l}: view/set the \code{short_label} slot of a \code{stype}
\code{\link{context}}.

\item \code{long_label_l}: view/set the \code{long_label} slot of a \code{stype}
\code{\link{context}}.

\item \code{description_l}: view/set the \code{description} slot of a \code{stype}
\code{\link{context}}.

\item \code{security_type_l}: view/set the \code{security_type} slot of a \code{stype}
\code{\link{context}}.

\item \code{purpose_l}: view/set the \code{purpose} slot of a \code{stype}
\code{\link{context}}.

\item \code{study_role_l}: view/set the \code{study_role} slot of a \code{stype}
\code{\link{context}}'s \code{\link{purpose}}.

\item \code{tags_l}: view/set the \code{tags} slot of a \code{stype}
\code{\link{context}}'s \code{\link{purpose}}.
}}

\keyword{datasets}
