% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{stype_predicates}
\alias{stype_predicates}
\alias{is_constant}
\alias{is_not_constant}
\title{Useful stype predicates}
\usage{
is_constant(x)

is_not_constant(x)
}
\arguments{
\item{x}{a \code{stype}}
}
\description{
Useful stype predicates
}
\section{Functions}{
\itemize{
\item \code{is_constant}: is this vector constant?

\item \code{is_not_constant}: is this vector not constant?
}}

