% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_binary.R
\name{v_binary}
\alias{v_binary}
\alias{new_binary}
\alias{is_binary}
\alias{as_binary}
\title{Count vectors}
\usage{
new_binary(x = logical(), desc = description())

v_binary(x = logical(), ...)

is_binary(x)

as_binary(x)
}
\description{
Some desc

constructor function for count objects
}
