% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_binary.R
\name{v_binary}
\alias{v_binary}
\alias{is_binary}
\alias{as_binary}
\alias{as.character.v_binary}
\alias{as_canonical.v_binary}
\title{Binary vectors}
\usage{
v_binary(
  x = ptypeFUN(),
  internal_name = "",
  context,
  extra_descriptors = list()
)

is_binary(x)

as_binary(x)

\method{as.character}{v_binary}(x, ...)

\method{as_canonical}{v_binary}(x)
}
\arguments{
\item{x}{a \code{logical} vector}

\item{internal_name}{the internal name of the variable}

\item{context}{a \code{\link{context}}}

\item{extra_descriptors}{A \code{list} of \code{\link{descriptors}} functions
appended to the default \code{\link{descriptors}}.}

\item{...}{passed to other methods such as \code{as.character}}
}
\description{
{
Support: \eqn{0, 1}{\{0, 1\}} (plus \code{\link{NA}})

Prototype: \code{\link{logical}}
}
}
\seealso{
Other stype types: 
\code{\link{tbl_analysis}},
\code{\link{v_character}},
\code{\link{v_continuous_nonneg}},
\code{\link{v_continuous}},
\code{\link{v_count}},
\code{\link{v_event_time}},
\code{\link{v_nominal}},
\code{\link{v_ordered}},
\code{\link{v_rcensored}}
}
\concept{stype types}
