% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_character.R
\name{v_character}
\alias{v_character}
\alias{chr}
\alias{is_character}
\alias{as_character}
\alias{as.character.v_character}
\alias{as_canonical.v_character}
\title{Character vectors}
\usage{
v_character(
  x = ptypeFUN(),
  internal_name = "",
  context,
  extra_descriptors = list()
)

chr(x = ptypeFUN(), internal_name = "", context, extra_descriptors = list())

is_character(x)

as_character(x)

\method{as.character}{v_character}(x, ...)

\method{as_canonical}{v_character}(x)
}
\arguments{
\item{x}{a \code{character} vector}

\item{internal_name}{the internal name of the variable}

\item{context}{a \code{\link{context}}}

\item{extra_descriptors}{A \code{list} of \code{\link{descriptors}} functions
appended to the default \code{\link{descriptors}}.}

\item{...}{passed to other methods such as \code{as.character}}
}
\description{
{
Support: see \code{\link{character}}

Prototype: \code{\link{character}}
}
}
\seealso{
Other stype types: 
\code{\link{tbl_analysis}},
\code{\link{v_binary}},
\code{\link{v_continuous_nonneg}},
\code{\link{v_continuous}},
\code{\link{v_count}},
\code{\link{v_event_time}},
\code{\link{v_nominal}},
\code{\link{v_ordered}},
\code{\link{v_rcensored}}
}
\concept{stype types}
