% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_character.R
\name{v_character}
\alias{v_character}
\alias{new_character}
\alias{is_character}
\alias{as_character}
\title{Character vectors}
\usage{
new_character(x = character(), .desc = description(),
  .context = context())

v_character(x = character(), context)

is_character(x)

as_character(x)
}
\arguments{
\item{x}{a \code{character} vector}

\item{.desc}{a \\code{\link{description}}}

\item{.context}{a \code{\link{context}}}

\item{context}{a \code{\link{context}}}
}
\description{
Some desc

constructor function for character objects
}
