% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_continuous.R
\name{v_continuous}
\alias{v_continuous}
\alias{is_continuous}
\alias{as_continuous}
\alias{as_canonical.v_continuous}
\title{Continuous vectors}
\usage{
v_continuous(x = double(), internal_name = "", context)

is_continuous(x)

as_continuous(x)

\method{as_canonical}{v_continuous}(x)
}
\arguments{
\item{x}{a \code{double} vector}

\item{internal_name}{the internal name of the variable}

\item{context}{a \code{\link{context}}}
}
\description{
{
Support: \eqn{\mathbf{R}}{the Reals}* (plus \code{\link{NA_real_}})

Prototype: \code{\link{double}}

* - i.e. floating-point number
}
}
\seealso{
Other stype types: 
\code{\link{v_binary}},
\code{\link{v_character}},
\code{\link{v_continuous_nonneg}},
\code{\link{v_count}},
\code{\link{v_event_time}},
\code{\link{v_nominal}},
\code{\link{v_ordered}},
\code{\link{v_rcensored}}
}
\concept{stype types}
