% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_continuous.R
\name{v_continuous}
\alias{v_continuous}
\alias{new_continuous}
\alias{is_continuous}
\alias{as_continuous}
\title{Continuous vectors}
\usage{
new_continuous(x = double(), .internal_name = character(),
  .data_summary = data_summary(), .context = context())

v_continuous(x = double(), internal_name = "", context)

is_continuous(x)

as_continuous(x)
}
\arguments{
\item{x}{a \code{double} vector}

\item{.context}{a \code{\link{context}}}

\item{context}{a \code{\link{context}}}
}
\description{
Some desc

constructor function for count objects
}
