% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_continuous.R
\name{v_continuous}
\alias{v_continuous}
\alias{is_continuous}
\alias{as_continuous}
\alias{as_canonical.v_continuous}
\title{Continuous vectors}
\usage{
v_continuous(x = double(), internal_name = "", context)

is_continuous(x)

as_continuous(x)

\method{as_canonical}{v_continuous}(x)
}
\arguments{
\item{x}{a \code{double} vector}

\item{internal_name}{the internal name of the variable}

\item{context}{a \code{\link{context}}}
}
\description{
Some desc

constructor function for count objects
}
