% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_continuous.R
\name{v_continuous}
\alias{v_continuous}
\alias{new_continuous}
\alias{is_continuous}
\alias{as_continuous}
\title{Continuous vectors}
\usage{
new_continuous(x = double(), desc = description())

v_continuous(x = double(), ...)

is_continuous(x)

as_continuous(x)
}
\description{
Some desc

constructor function for count objects
}
