% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_continuous_nonneg.R
\name{v_continuous_nonneg}
\alias{v_continuous_nonneg}
\alias{is_continuous_nonneg}
\alias{is_nonneg}
\alias{as_nonneg_continuous}
\alias{as_canonical.v_continuous_nonneg}
\title{Non-negative continuous vectors}
\usage{
v_continuous_nonneg(x = double(), internal_name = "", context)

is_continuous_nonneg(x)

is_nonneg(x)

as_nonneg_continuous(x)

\method{as_canonical}{v_continuous_nonneg}(x)
}
\arguments{
\item{x}{nonnegative \code{double}}

\item{internal_name}{the internal name of the variable}

\item{context}{a \code{\link{context}}}
}
\description{
{
Support: \eqn{\mathbf{R^+}}{the nonnegative Reals}* (plus \code{\link{NA_real_}})

Prototype: \code{\link{double}}

* - technically a double-precision number
}
}
\seealso{
Other stype types: 
\code{\link{v_binary}},
\code{\link{v_character}},
\code{\link{v_continuous}},
\code{\link{v_count}},
\code{\link{v_event_time}},
\code{\link{v_nominal}},
\code{\link{v_ordered}},
\code{\link{v_rcensored}}
}
\concept{stype types}
