% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_continuous_nonneg.R
\name{v_continuous_nonneg}
\alias{v_continuous_nonneg}
\alias{is_continuous_nonneg}
\alias{is_nonneg}
\alias{as_nonneg_continuous}
\alias{as_canonical.v_continuous_nonneg}
\title{continuous_nonneg vectors}
\usage{
v_continuous_nonneg(x = double(), internal_name = "", context)

is_continuous_nonneg(x)

is_nonneg(x)

as_nonneg_continuous(x)

\method{as_canonical}{v_continuous_nonneg}(x)
}
\arguments{
\item{x}{nonnegative \code{double}}

\item{internal_name}{the internal name of the variable}

\item{context}{a \code{\link{context}}}
}
\description{
Some desc

constructor function for count objects
}
