% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_count.R
\name{v_count}
\alias{v_count}
\alias{is_count}
\alias{as_count}
\alias{as_canonical}
\alias{as_canonical.v_count}
\title{Count vectors}
\usage{

v_count(x = integer(), internal_name = "", context)

is_count(x)

as_count(x)

as_canonical(x)

\method{as_canonical}{v_count}(x)
}
\arguments{
\item{x}{a \code{integer} vector}

\item{internal_name}{the internal name of the variable}

\item{context}{a \code{\link{context}}}
}
\description{
Some desc

constructor function for count objects
}
