% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_count.R
\name{v_count}
\alias{v_count}
\alias{new_count}
\alias{is_count}
\alias{as_count}
\title{Count vectors}
\usage{
new_count(x = integer(), desc = description())

v_count(x = integer(), ...)

is_count(x)

as_count(x)
}
\description{
Some desc

constructor function for count objects
}
