% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_count.R
\name{v_count}
\alias{v_count}
\alias{new_count}
\alias{is_count}
\alias{as_count}
\title{Count vectors}
\usage{
new_count(x = integer(), .internal_name = character(),
  .data_summary = data_summary(), .context = context())

v_count(x = integer(), internal_name = "", context)

is_count(x)

as_count(x)
}
\arguments{
\item{x}{a \code{integer} vector}

\item{.context}{a \code{\link{context}}}

\item{context}{a \code{\link{context}}}

\item{.desc}{a \\code{\link{description}}}
}
\description{
Some desc

constructor function for count objects
}
