% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_event_time.R
\name{v_event_time}
\alias{v_event_time}
\alias{is_event_time}
\alias{as_time_to_event}
\alias{as_event_time}
\alias{as_canonical.v_event_time}
\title{Time to event vectors}
\usage{
v_event_time(
  x = ptypeFUN(),
  internal_name = "",
  context,
  extra_descriptors = list()
)

is_event_time(x)

as_time_to_event(x)

as_event_time(x)

\method{as_canonical}{v_event_time}(x)
}
\arguments{
\item{x}{vector of values}

\item{internal_name}{the internal name of the variable}

\item{context}{a \code{\link{context}}}

\item{extra_descriptors}{A \code{list} of \code{\link{descriptors}} functions
appended to the default \code{\link{descriptors}}.}
}
\description{
{
Support: \eqn{\mathbf{R^+}}{the nonnegative Reals}* (plus \code{\link{NA_real_}})

Prototype: \code{\link{double}}

* - i.e. floating-point number
}
}
\seealso{
Other stype types: 
\code{\link{v_binary}},
\code{\link{v_character}},
\code{\link{v_continuous_nonneg}},
\code{\link{v_continuous}},
\code{\link{v_count}},
\code{\link{v_nominal}},
\code{\link{v_ordered}},
\code{\link{v_rcensored}}
}
\concept{stype types}
