% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_event_time.R
\name{v_event_time}
\alias{v_event_time}
\alias{is_event_time}
\alias{as_time_to_event}
\alias{as_event_time}
\alias{as_canonical.v_event_time}
\title{event_time vectors}
\usage{
v_event_time(x = v_continuous_nonneg(), internal_name = "", context)

is_event_time(x)

as_time_to_event(x)

as_event_time(x)

\method{as_canonical}{v_event_time}(x)
}
\arguments{
\item{x}{vector of values}

\item{internal_name}{the internal name of the variable}

\item{context}{a \code{\link{context}}}
}
\description{
Some desc

constructor function for time to evet objects
}
