% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_nominal.R
\name{v_nominal}
\alias{v_nominal}
\alias{new_nominal}
\alias{is_nominal}
\alias{as_canonical.v_nominal}
\title{v_nominal S3 class}
\usage{
new_nominal(
  x = integer(),
  .levels = character(),
  .internal_name = character(),
  .data_summary = data_summary(),
  .context = context()
)

v_nominal(x = factor(), internal_name = "", context)

is_nominal(x)

\method{as_canonical}{v_nominal}(x)
}
\arguments{
\item{x}{a \code{factor}}

\item{.levels}{Character vector of labels.}

\item{.internal_name}{the internal name of the variable}

\item{.data_summary}{a \code{\link{data_summary}}}

\item{.context}{a \code{\link{context}}}

\item{internal_name}{the internal name of the variable}

\item{context}{a \code{\link{context}}}
}
\description{
A nominal (categorial) variable
}
