% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_nominal.R
\name{v_nominal}
\alias{v_nominal}
\alias{is_nominal}
\alias{as_canonical.v_nominal}
\title{v_nominal S3 class}
\usage{

v_nominal(x = factor(), internal_name = "", context)

is_nominal(x)

\method{as_canonical}{v_nominal}(x)
}
\arguments{
\item{x}{a \code{factor}}

\item{internal_name}{the internal name of the variable}

\item{context}{a \code{\link{context}}}
}
\description{
A nominal (categorial) variable
}
