% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_nominal.R
\name{v_nominal}
\alias{v_nominal}
\alias{new_nominal}
\alias{v_nomimal}
\title{v_nominal S3 class}
\usage{
new_nominal(x = integer(), .levels = character(),
  .desc = description(), .context = context())

v_nomimal(x = factor(), context)
}
\arguments{
\item{x}{a \code{factor}}

\item{.levels}{Character vector of labels.}

\item{.desc}{a \\code{\link{description}}}

\item{.context}{a \code{\link{context}}}

\item{context}{a \code{\link{context}}}
}
\description{
A nominal (categorial) variable
}
