% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_ordered.R
\name{v_ordered}
\alias{v_ordered}
\alias{is_ordered}
\alias{as_canonical.v_ordered}
\title{v_ordered S3 class}
\usage{
v_ordered(x = factor(ordered = TRUE), internal_name = "", context)

is_ordered(x)

\method{as_canonical}{v_ordered}(x)
}
\arguments{
\item{x}{a \code{factor}}

\item{internal_name}{the internal name of the variable}

\item{context}{a \code{\link{context}}}
}
\description{
An ordered (categorial) variable
}
