% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_ordered.R
\name{v_ordered}
\alias{v_ordered}
\alias{is_ordered}
\alias{as_canonical.v_ordered}
\title{Ordered categorical vectors}
\usage{
v_ordered(
  x = factor(ordered = TRUE),
  internal_name = "",
  context,
  extra_descriptors = list()
)

is_ordered(x)

\method{as_canonical}{v_ordered}(x)
}
\arguments{
\item{x}{a \code{factor}}

\item{internal_name}{the internal name of the variable}

\item{context}{a \code{\link{context}}}

\item{extra_descriptors}{A \code{list} of \code{\link{descriptors}} functions
appended to the default \code{\link{descriptors}}.}
}
\description{
{
Support: see \code{\link{factor}}

Prototype: \code{\link{integer}}
}
}
\seealso{
Other stype types: 
\code{\link{tbl_analysis}},
\code{\link{v_binary}},
\code{\link{v_character}},
\code{\link{v_continuous_nonneg}},
\code{\link{v_continuous}},
\code{\link{v_count}},
\code{\link{v_event_time}},
\code{\link{v_nominal}},
\code{\link{v_rcensored}}
}
\concept{stype types}
