% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_rcensored.R
\name{v_rcensored}
\alias{v_rcensored}
\alias{rcen}
\alias{is_rcensored}
\alias{as.character.v_rcensored}
\alias{as_canonical.v_rcensored}
\title{Right censored vectors}
\usage{
v_rcensored(
  outcomes = list(),
  censors,
  end_time = Inf,
  internal_name = "",
  context,
  extra_descriptors = list()
)

rcen(
  outcomes = list(),
  censors,
  end_time = Inf,
  internal_name = "",
  context,
  extra_descriptors = list()
)

is_rcensored(x)

\method{as.character}{v_rcensored}(x, ...)

\method{as_canonical}{v_rcensored}(x)
}
\arguments{
\item{outcomes}{A \code{list} of \code{v_event_time} vectors that define the
outcomes. The order of this list defines the precedence of outcomes. That is,
if the first outcome and second outcome occur at the same time, the first
outcome is the reason for the outcome.}

\item{censors}{A \code{list} of \code{v_event_time} vectors that define the
censor The order of this list defines the precedence of censoring That is,
if the first censor and second censor occur at the same time, the first
censor is the reason for the censoring}

\item{end_time}{A \code{numeric} scalar defining the end of follow-up.}

\item{internal_name}{the internal name of the variable}

\item{context}{a \code{\link{context}}}

\item{extra_descriptors}{A \code{list} of \code{\link{descriptors}} functions
appended to the default \code{\link{descriptors}}.}

\item{x}{a \code{integer} vector}

\item{...}{passed to other methods such as \code{as.character}}
}
\description{
\code{v_rcensored} provides a right-censored vector. Unlike other representations,
such the \code{survival} package's \code{Surv} object, \code{v_rcensored} can be subset with
\code{[} and concatenated with \code{c} as you would any other vector. The type is
implemented as a \code{\link[vctrs]{new_rcrd}} where the necessary data are
contained in \code{\link[vctrs]{fields}}.
}
\details{
\code{as_canonical} casts the vector to a \code{list}. See \code{\link{v_rcensored_accessors}}
for functions to access components of a \code{v_rcensored}.

When printed, an open right triangle indicates an observation was censored.
A closed right triangle indicates an observation reached \code{end_time} without
being censored or having an outcome. No triangles indicated an observation
that had one of the outcomes.
}
\seealso{
Other stype types: 
\code{\link{tbl_analysis}},
\code{\link{v_binary}},
\code{\link{v_character}},
\code{\link{v_continuous_nonneg}},
\code{\link{v_continuous}},
\code{\link{v_count}},
\code{\link{v_event_time}},
\code{\link{v_nominal}},
\code{\link{v_ordered}}
}
\concept{stype types}
