% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_rcensored.R
\name{v_rcensored}
\alias{v_rcensored}
\alias{is_rcensored}
\alias{as_canonical.v_rcensored}
\title{Right censored vectors}
\usage{
v_rcensored(
  outcomes = list(),
  censors = list(),
  end_time = Inf,
  internal_name = "",
  context
)

is_rcensored(x)

\method{as_canonical}{v_rcensored}(x)
}
\arguments{
\item{outcomes}{a \code{list} of \code{v_event_time} vectors}

\item{censors}{a \code{list} of \code{v_event_time} vectors}

\item{end_time}{a scalar}

\item{internal_name}{the internal name of the variable}

\item{context}{a \code{\link{context}}}

\item{x}{a \code{integer} vector}
}
\description{
Some desc

constructor function for right censored objects
}
