% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_rcensored.R
\name{v_rcensored}
\alias{v_rcensored}
\alias{is_rcensored}
\alias{as_canonical.v_rcensored}
\title{Right censored vectors}
\usage{
v_rcensored(
  outcomes = list(),
  censors = list(),
  end_time = Inf,
  internal_name = "",
  context
)

is_rcensored(x)

\method{as_canonical}{v_rcensored}(x)
}
\arguments{
\item{outcomes}{A \code{list} of \code{v_event_time} vectors that define the 
outcomes. The order of this list defines the precedence of outcomes. That is,
if the first outcome and second outcome occur at the same time, the first 
outcome is the reason for the outcome.}

\item{censors}{A \code{list} of \code{v_event_time} vectors that define the 
censor The order of this list defines the precedence of censoring That is,
if the first censor and second censor occur at the same time, the first 
censor is the reason for the censoring}

\item{end_time}{A \code{numeric} scalar defining the end of follow-up.}

\item{internal_name}{the internal name of the variable}

\item{context}{a \code{\link{context}}}

\item{x}{a \code{integer} vector}
}
\description{
TODO: add a description
}
\seealso{
Other stype types: 
\code{\link{v_binary}},
\code{\link{v_character}},
\code{\link{v_continuous_nonneg}},
\code{\link{v_continuous}},
\code{\link{v_count}},
\code{\link{v_event_time}},
\code{\link{v_nominal}},
\code{\link{v_ordered}}
}
\concept{stype types}
