% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_count.R, R/v_continuous.R,
%   R/v_continuous_nonneg.R, R/v_binary.R
\name{vec_arith}
\alias{vec_arith}
\alias{vec_arith.v_count}
\alias{vec_arith.integer}
\alias{vec_math.v_count}
\alias{vec_arith.v_continuous}
\alias{vec_arith.double}
\alias{vec_math.v_continuous}
\alias{vec_type2.numeric}
\alias{vec_arith.v_binary}
\alias{vec_math.v_binary}
\title{Arithmetic ops}
\usage{
\method{vec_arith}{v_count}(op, x, y)

\method{vec_arith}{integer}(op, x, y)

\method{vec_math}{v_count}(fun, x, ...)

\method{vec_arith}{v_continuous}(op, x, y)

\method{vec_arith}{double}(op, x, y)

\method{vec_math}{v_continuous}(fun, x, ...)

\method{vec_type2}{numeric}(x, y, ...)

\method{vec_arith}{v_binary}(op, x, y)

\method{vec_math}{v_binary}(fun, x, ...)
}
\arguments{
\item{op}{An arithmetic operator as a string}

\item{x}{A pair of vectors. For \code{!}, unary \code{+} and unary \code{-}, \code{y} will be
a sentinel object of class \code{MISSING}, as created by \code{MISSING()}.}

\item{y}{A pair of vectors. For \code{!}, unary \code{+} and unary \code{-}, \code{y} will be
a sentinel object of class \code{MISSING}, as created by \code{MISSING()}.}

\item{fun}{An mathematical function as a string}

\item{...}{An additional arguments.}
}
\description{
Arithmetic ops
}
