% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_count.R, R/v_continuous.R, R/v_binary.R
\name{vec_math}
\alias{vec_math}
\alias{vec_math.v_count}
\alias{vec_math.v_continuous}
\alias{vec_math.v_binary}
\title{Math}
\usage{
\method{vec_math}{v_count}(fun, x, ...)

\method{vec_math}{v_continuous}(fun, x, ...)

\method{vec_math}{v_binary}(fun, x, ...)
}
\arguments{
\item{fun}{fun}

\item{x}{x}

\item{...}{dots}
}
\description{
Math
}
